+++ lib/transformer/nestedComment.ts
@@ -1,10 +1,8 @@
-import { User, Record } from "@prisma/client";
 import { Comment } from "../../@types/Comment";
 /**
  * A public function that creates a nested structure of comments based on the complete array of comments.
  * It looks for a parentId in the comment before recursively calling itself to find the children of the comment.
  */
-
 export default function createNestedStructure(
   comments: Comment[],
   parentId: string | null = null
