+++ lib/transformer/profileTransformer.ts
@@ -5,8 +5,6 @@ export class ProfileTransformer {
    * A public method to transform the user from the database to strip any information we don't want sent to the client.
    * It also injects the access token into the payload
    * @param idpUser
-   * @param accessToken
-   * @returns
    */
   public static transformProfile(idpUser: IdpUser): User {
     return {
