--- services/comments/comment.dto.ts
@@ -1,4 +1,0 @@
-export class NewComment {
-  message: string;
-  emailNotify: boolean;
-}
